refactor(base): flat module structure
