+++ packages/base/src/form/cascader.ts
@@ -1,4 +1,4 @@
-import type { PopupFormItem } from '../popup'
+import type { PopupFormItem } from './popup'
 
 /**
  * 级联选择表单项类型
@@ -63,4 +63,36 @@ export type CascaderFormItemOption = {
    * 子选项
    */
   children?: CascaderFormItemOption[]
+}
+
+export class CascaderHelper {
+  /**
+   * 根据 value 值获取其在级联选项中的的路径选项（数组）
+   */
+  static getOptionsPathByValue = (
+    options: CascaderFormItemOption[],
+    value?: CascaderFormItemOption['value'],
+  ): CascaderFormItemOption[] => {
+    if (!value || !options.length) return []
+
+    const _getOptionsPathByValue = (
+      options: CascaderFormItemOption[],
+    ): CascaderFormItemOption[] | undefined => {
+      for (const option of options) {
+        if (option.value === value) {
+          return [option]
+        }
+
+        if (option.children) {
+          const selectedOptions = _getOptionsPathByValue(option.children)
+          if (selectedOptions) {
+            return [option, ...selectedOptions]
+          }
+        }
+      }
+      return undefined
+    }
+
+    return _getOptionsPathByValue(options) ?? []
+  }
 }
