--- packages/base/src/form/cascader/index.ts
@@ -1,2 +1,0 @@
-export * from './CascaderFormItem'
-export * from './helper'
