--- packages/base/src/form/checkbox/index.ts
@@ -1,1 +1,0 @@
-export * from './CheckboxFormItem'
