--- packages/base/src/form/custom/index.ts
@@ -1,1 +1,0 @@
-export * from './CustomFormItem'
