--- packages/base/src/form/date/index.ts
@@ -1,1 +1,0 @@
-export * from './DateFormItem'
