+++ packages/base/src/form/input.ts
@@ -1,4 +1,4 @@
-import type { BaseFormItem } from '../base'
+import type { BaseFormItem } from './base'
 
 /**
  * 输入表单项类型
