--- packages/base/src/form/input/index.ts
@@ -1,1 +1,0 @@
-export * from './InputFormItem'
