+++ packages/base/src/form/popup.ts
@@ -1,4 +1,4 @@
-import type { BaseFormItem } from '../base'
+import type { BaseFormItem } from './base'
 
 /**
  * 弹出层表单项类型
