--- packages/base/src/form/popup/index.ts
@@ -1,1 +1,0 @@
-export * from './PopupFormItem'
