--- packages/base/src/form/radio/helper.ts
@@ -1,11 +1,0 @@
-import { log } from '@cphayim-enc/shared'
-
-import type { RadioFormItem } from './RadioFormItem'
-
-export class RadioHelper {
-  static verifyRadioFormItem(item: RadioFormItem) {
-    if (!item.radioOptions || !item.radioOptions.length) {
-      log.warn(`RadioFormItem.radioOptions is empty`)
-    }
-  }
-}
