--- packages/base/src/form/radio/index.ts
@@ -1,1 +1,0 @@
-export * from './RadioFormItem'
