--- packages/base/src/form/rate/index.ts
@@ -1,1 +1,0 @@
-export * from './RateFormItem'
