--- packages/base/src/form/select/index.ts
@@ -1,1 +1,0 @@
-export * from './SelectFormItem'
