--- packages/base/src/form/switch/index.ts
@@ -1,1 +1,0 @@
-export * from './SwitchFormItem'
