--- packages/base/src/form/time/index.ts
@@ -1,1 +1,0 @@
-export * from './TimeFormItem'
