--- packages/base/src/form/upload/index.ts
@@ -1,2 +1,0 @@
-export * from './UploadFormItem'
-export * from './helper'
